+++ Assets/BossRepositionWeakspot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5eeb96102065347958a9f66bbe8f3a0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
