+++ Assets/Prefabs/Moon Rock Low.prefab
@@ -11,8 +11,8 @@ GameObject:
   - component: {fileID: 2846665734829257844}
   - component: {fileID: 4775774813234597603}
   - component: {fileID: 7375655881103720199}
-  - component: {fileID: 7535323662806828595}
-  m_Layer: 0
+  - component: {fileID: 1972665360369244947}
+  m_Layer: 10
   m_Name: Moon Rock Low
   m_TagString: Obstacle
   m_Icon: {fileID: 0}
@@ -32,6 +32,7 @@ Transform:
   m_ConstrainProportionsScale: 1
   m_Children:
   - {fileID: 5945742139243832024}
+  - {fileID: 1043242006199398446}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -64,13 +65,59 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   explosionPrefab: {fileID: 180128, guid: d64dadb073f5bb64f97af17f6be6e028, type: 3}
---- !u!65 &7535323662806828595
-BoxCollider:
+--- !u!64 &1972665360369244947
+MeshCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3227589770988646094}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_Enabled: 1
+  serializedVersion: 4
+  m_Convex: 1
+  m_CookingOptions: 30
+  m_Mesh: {fileID: 1242625235393168646, guid: 50e1f478aad4c8b419dcd4e86fb2f213, type: 3}
+--- !u!1 &5063520386209560292
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1043242006199398446}
+  - component: {fileID: 6264896897041231533}
+  m_Layer: 11
+  m_Name: BoxCollider
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1043242006199398446
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5063520386209560292}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 2846665734829257844}
+  m_RootOrder: 1
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!65 &6264896897041231533
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5063520386209560292}
   m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_Enabled: 1
@@ -88,7 +135,7 @@ GameObject:
   - component: {fileID: 5945742139243832024}
   - component: {fileID: 267325369918777666}
   - component: {fileID: 5142025630297245844}
-  m_Layer: 0
+  m_Layer: 10
   m_Name: Low_Poly_Retopo001
   m_TagString: Untagged
   m_Icon: {fileID: 0}
