+++ Assets/Scripts/Boss/BossControl.cs
@@ -34,7 +34,6 @@ public class BossControl : MonoBehaviour
 
     [Header("Boss weak spot variables")]
     [SerializeField] GameObject weakSpotPf;
-    MeshCollider meshCollider;
     Mesh bossMesh;
     Camera mainCamera;
     MeshFilter meshFilter;
@@ -50,10 +49,6 @@ public class BossControl : MonoBehaviour
     [Header("Audio Settings")]
     [SerializeField] public AudioClip weakSpotSound;
     [SerializeField][Range(0f, 1f)] public float weakSpotVolume = 0.7f;
-
-    // Set up for weak spot generation check
-    GameObject[] weakspots = new GameObject[3];
-    bool[] isColliding = new bool[3];   // Flag to identify if mesh collider is needed
 
     // Start is called before the first frame update
     void Awake()
@@ -78,8 +73,6 @@ public class BossControl : MonoBehaviour
         meshFilter = gameObject.GetComponent<MeshFilter>();
         bossMesh = meshFilter.mesh;
         occlusionMask = gameObject.layer;
-        meshCollider = gameObject.GetComponent<MeshCollider>();
-        meshCollider.enabled = true;
         WeakSpotStatCol = new Dictionary<int, Color>()
         {
             { 0, WeakSpotCols0 },
@@ -125,8 +118,6 @@ public class BossControl : MonoBehaviour
         // Finish condition for one shooting interval
         if (bossDead)
             StopShooting();
-
-        gameObject.GetComponent<MeshCollider>().enabled = isColliding[0] || isColliding[1] || isColliding[2];
     }
 
     // Shoot one carrot
@@ -194,7 +185,6 @@ public class BossControl : MonoBehaviour
     {
         if (!bossDead)
         {
-            meshCollider.enabled = false;
             bossDead = true;
             RemoveAllWeakSpots();
 
@@ -322,12 +312,9 @@ public class BossControl : MonoBehaviour
                 Vector3 normal = bossMesh.normals[closestVertexIndex];
                 Vector3 worldNormal = meshFilter.transform.TransformDirection(normal);
 
-                weakspots[chosenIndices.Count - 1] = Instantiate(weakSpotPf, randomPoint, Quaternion.LookRotation(worldNormal), gameObject.transform);
-                isColliding[chosenIndices.Count - 1] = true;
+                Instantiate(weakSpotPf, randomPoint, Quaternion.LookRotation(worldNormal), gameObject.transform);
             }
         }
-
-        gameObject.GetComponent<MeshCollider>().enabled = true;
     }
 
     // To find normal (Called by GetWeakSpots())
@@ -348,38 +335,6 @@ public class BossControl : MonoBehaviour
         }
 
         return closestIndex;
-    }
-
-    // Make sure weakspot is not buried under boss mesh (Called on collision btwn boss and weak spot)
-    void AdjustWeakSpot(GameObject weakSpot)
-    {
-        weakSpot.transform.localPosition += (Vector3.forward * 0.1f);
-    }
-
-    private void OnCollisionStay(Collision collision)
-    {
-        // Make sure weakspot is not buried under boss mesh
-        if (collision.gameObject.CompareTag("WeakSpot"))
-        {
-            for (int i = 0; i < 3; i++)
-            {
-                if (weakspots[i] == collision.gameObject)
-                    isColliding[i] = true;
-            }
-            AdjustWeakSpot(collision.gameObject);
-        }
-    }
-
-    private void OnCollisionExit(Collision collision)
-    {
-        if (collision.gameObject.CompareTag("WeakSpot"))
-        {
-            for (int i = 0; i < 3; i++)
-            {
-                if (weakspots[i] == collision.gameObject)
-                    isColliding[i] = false;
-            }
-        }
     }
 
     // Call to change color of weak spot on attack (collision between laser and weakspot)
